feat: implement useTodoList hook for managing TodoList state and progress
